requirements.txt
autogen
jsonlines==3.1.0
matplotlib==3.7.1
numpy==1.24.3
openai==1.7.2
pandas==2.1.4
pyautogen==0.2.0
python_Levenshtein==0.23.0
seaborn==0.13.1
termcolor==2.4.0
wolframalpha==5.0.0
